added config var for enabling latency metriccs
