+++ src/k2/cmd/nodepool/nodepool_main.cpp
@@ -39,7 +39,8 @@ int main(int argc, char** argv) {
         ("k23si_cpo_endpoint", bpo::value<k2::String>(), "the endpoint for k2 CPO service")
         ("k23si_query_pagination_limit", bpo::value<uint32_t>(), "Max records to return in a single query response")
         ("k23si_query_push_limit", bpo::value<uint32_t>(), "Min records in response needed to avoid a push during query processing")
-        ("k23si_persistence_endpoints", bpo::value<std::vector<k2::String>>()->multitoken()->default_value(std::vector<k2::String>()), "A space-delimited list of k2 persistence endpoints, each core will pick one endpoint");
+        ("k23si_persistence_endpoints", bpo::value<std::vector<k2::String>>()->multitoken()->default_value(std::vector<k2::String>()), "A space-delimited list of k2 persistence endpoints, each core will pick one endpoint")
+        ("enable_metrics", bpo::value<bool>(), "For enabling latency metrics");
 
     app.addApplet<k2::APIServer>();
     app.addApplet<k2::TSO_ClientLib>();
