+++ src/k2/module/k23si/Module.h
@@ -41,7 +41,6 @@ Copyright(c) 2020 Futurewei Cloud
 #include "Config.h"
 #include "Persistence.h"
 #include "Log.h"
-#include "utils.h"
 
 namespace k2 {
