+++ src/k2/module/k23si/Persistence.cpp
@@ -51,6 +51,7 @@ void Persistence::_registerMetrics() {
     labels.push_back(sm::label_instance("total_cores", seastar::smp::count));
 
     _metric_groups.add_group("Nodepool", {
+        sm::make_counter("flushes", _flushes, sm::description("Number of flushes"), labels),
         sm::make_histogram("flush_latency", [this]{ return _flushLatency.getHistogram();},
                 sm::description("Latency of Persistence Flush"), labels)
     });
@@ -77,7 +78,7 @@ seastar::future<> Persistence::stop() {
 }
 
 seastar::future<Status> Persistence::flush() {
-    auto start = k2::Clock::now();
+    k2::OperationLatencyReporter reporter(_flushes, _flushLatency); // for reporting metrics
     ++_flushId;
     K2LOG_D(log::skvsvr, "flush with bs={}, proms={}, fid={}", (_buffer? _buffer->getSize() : 0), _pendingProms.size(), _flushId);
     if (!_buffer) {
@@ -121,10 +122,8 @@ seastar::future<Status> Persistence::flush() {
             return seastar::make_ready_future<Status>(std::move(status));
         });
     return _chainFlushResponse()
-            .then([this,start](auto&& response){
-                    auto end = k2::Clock::now();
-                    auto dur = end - start;
-                    _flushLatency.add(dur);
+            .then([this, reporter=std::move(reporter)](auto&& response) mutable{
+                    reporter.report();
                     return std::move(response);
             });
 }
